+++ src/main/java/TechTigers/BicycleBuddies/models/AbstractEntity.java
@@ -11,13 +11,13 @@ import java.util.Objects;
 public abstract class AbstractEntity {
     @Id
     @GeneratedValue
-    private int id;
+    private Long id;
 
-    public long getId(){
+    public Long getId(){
         return id;
     }
 
-    public void setId(int id) {this.id = id;}
+    public void setId(Long id) {this.id = id;}
 
     @Override
     public boolean equals(Object o) {
